Open Meet chat window only upon pressing down the configured key
Something has changed in Google Meet such that the chat isn't marked as having focus quickly enough, so the key release event triggers a second click on the chat button. The effect is that the chat disappears almost immediately after being opened.

Therefore, now the chat button is not clicked upon the key release event.